fix get_all

+++ src/realm/array_direct.hpp
@@ -258,16 +258,17 @@ public:
     bf_iterator(bf_iterator&&) = default;
     bf_iterator& operator=(const bf_iterator&) = default;
     bf_iterator& operator=(bf_iterator&&) = default;
-    bf_iterator(const uint64_t* data_area, size_t initial_offset, size_t field_size, size_t step_size, size_t index)
+    bf_iterator(const uint64_t* data_area, size_t initial_offset, uint8_t field_size, uint8_t step_size, size_t index)
     {
         init(data_area, initial_offset, field_size, step_size, index);
     }
 
-    inline void init(const uint64_t* data_area, size_t initial_offset, size_t field_size, size_t step_size, size_t index)
+    inline void init(const uint64_t* data_area, size_t initial_offset, uint8_t field_size, uint8_t step_size,
+                     size_t index)
     {
         this->data_area = (uint64_t*)data_area;
-        this->field_size = static_cast<uint8_t>(field_size);
-        this->step_size = static_cast<uint8_t>(step_size);
+        this->field_size = field_size;
+        this->step_size = step_size;
         this->offset = initial_offset;
         if (field_size < 64)
             mask = (1ULL << field_size) - 1;
@@ -509,7 +510,7 @@ bool inline any_field_NE(int width, uint64_t A, uint64_t B)
 
 // Populate all fields in a vector with a given value of a give width.
 // Bits outside of the given field are ignored.
-inline constexpr uint64_t populate(size_t width, uint64_t value)
+inline constexpr uint64_t populate(uint8_t width, uint64_t value)
 {
     value &= 0xFFFFFFFFFFFFFFFFULL >> (64 - width);
     if (width < 8) {
