+++ src/realm/integer_flex_compressor.cpp
@@ -30,12 +30,11 @@
 
 using namespace realm;
 
-void FlexCompressor::init_array(char* h, uint8_t flags, size_t v_width, size_t ndx_width, size_t v_size,
+void FlexCompressor::init_array(char* h, uint8_t flags, uint8_t v_width, uint8_t ndx_width, size_t v_size,
                                 size_t ndx_size) const
 {
     using Encoding = NodeHeader::Encoding;
-    init_header(h, Encoding::Flex, flags, static_cast<uint8_t>(v_width), static_cast<uint8_t>(ndx_width), v_size,
-                ndx_size);
+    init_header(h, Encoding::Flex, flags, v_width, ndx_width, v_size, ndx_size);
 }
 
 void FlexCompressor::copy_data(const Array& arr, const std::vector<int64_t>& values,
